fix risedev
