+++ risedev.yml
@@ -489,18 +489,16 @@ profile:
       - use: etcd
         unsafe-no-fsync: true
       - use: meta-node
+        enable-in-memory-kv-state-backend: true
       - use: compute-node
         port: 5687
         exporter-port: 1222
-        enable-in-memory-kv-state-backend: true
       - use: compute-node
         port: 5688
         exporter-port: 1223
-        enable-in-memory-kv-state-backend: true
       - use: compute-node
         port: 5689
         exporter-port: 1224
-        enable-in-memory-kv-state-backend: true
       - use: frontend
         port: 4565
         exporter-port: 2222
@@ -770,17 +768,8 @@ template:
     # Jaeger used by this compute node
     provide-jaeger: "jaeger*"
 
-    # Sanity check: should use shared storage if there're multiple compute nodes
-    provide-compute-node: "compute-node*"
-
-    # Sanity check: should start at lease one compactor if using shared object store
-    provide-compactor: "compactor*"
-
     # If `user-managed` is true, this service will be started by user with the above config
     user-managed: false
-
-    # Whether to enable in-memory pure KV state backend
-    enable-in-memory-kv-state-backend: false
 
     # Total available memory for the compute node in bytes
     total-memory-bytes: 8589934592
@@ -825,6 +814,24 @@ template:
 
     # Prometheus nodes used by dashboard service
     provide-prometheus: "prometheus*"
+
+    # Sanity check: should use shared storage if there're multiple compute nodes
+    provide-compute-node: "compute-node*"
+
+    # Sanity check: should start at lease one compactor if using shared object store
+    provide-compactor: "compactor*"
+
+    # Minio instances used by the cluster
+    provide-minio: "minio*"
+
+    # OpenDAL storage backend used by the cluster
+    provide-opendal: "opendal*"
+
+    # AWS s3 bucket used by the cluster
+    provide-aws-s3: "aws-s3*"
+
+    # Whether to enable in-memory pure KV state backend
+    enable-in-memory-kv-state-backend: false
 
   prometheus:
     # Advertise address of Prometheus
@@ -910,13 +917,8 @@ template:
     # Id of this instance
     id: compactor-${port}
 
-    # Minio instances used by this compute node
+    # Minio instances used by this compactor
     provide-minio: "minio*"
-
-    # OpenDAL storage backend used by this compute node
-    provide-opendal: "opendal*"
-    # AWS s3 bucket used by this compute node
-    provide-aws-s3: "aws-s3*"
 
     # Meta-nodes used by this compute node
     provide-meta-node: "meta-node*"
