+++ src/risedevtool/src/task/compactor_service.rs
@@ -17,13 +17,10 @@ use std::io::Write;
 use std::path::Path;
 use std::process::Command;
 
-use anyhow::{anyhow, Result};
+use anyhow::Result;
 
 use crate::util::{get_program_args, get_program_env_cmd, get_program_name};
-use crate::{
-    add_meta_node, add_storage_backend, CompactorConfig, ExecuteContext, HummockInMemoryStrategy,
-    Task,
-};
+use crate::{add_meta_node, CompactorConfig, ExecuteContext, Task};
 
 pub struct CompactorService {
     config: CompactorConfig,
@@ -45,19 +42,7 @@ impl CompactorService {
     }
 
     /// Apply command args according to config
-    pub fn apply_command_args(
-        cmd: &mut Command,
-        config: &CompactorConfig,
-        hummock_in_memory_strategy: HummockInMemoryStrategy,
-    ) -> Result<()> {
-        if matches!(
-            hummock_in_memory_strategy,
-            HummockInMemoryStrategy::Isolated
-        ) {
-            return Err(anyhow!(
-                "compactor cannot use in-memory hummock if remote object store is not provided"
-            ));
-        }
+    pub fn apply_command_args(cmd: &mut Command, config: &CompactorConfig) -> Result<()> {
         cmd.arg("--listen-addr")
             .arg(format!("{}:{}", config.listen_address, config.port))
             .arg("--prometheus-listener-addr")
@@ -77,18 +62,6 @@ impl CompactorService {
             cmd.arg("--compaction-worker-threads-number")
                 .arg(format!("{}", compaction_worker_threads_number));
         }
-
-        let provide_minio = config.provide_minio.as_ref().unwrap();
-        let provide_aws_s3 = config.provide_aws_s3.as_ref().unwrap();
-        let provide_opendal = config.provide_opendal.as_ref().unwrap();
-        add_storage_backend(
-            &config.id,
-            provide_opendal,
-            provide_minio,
-            provide_aws_s3,
-            hummock_in_memory_strategy,
-            cmd,
-        )?;
 
         let provide_meta_node = config.provide_meta_node.as_ref().unwrap();
         add_meta_node(provide_meta_node, cmd)?;
@@ -124,7 +97,7 @@ impl Task for CompactorService {
 
         cmd.arg("--config-path")
             .arg(Path::new(&prefix_config).join("risingwave.toml"));
-        Self::apply_command_args(&mut cmd, &self.config, HummockInMemoryStrategy::Disallowed)?;
+        Self::apply_command_args(&mut cmd, &self.config)?;
 
         if !self.config.user_managed {
             ctx.run_command(ctx.tmux_run(cmd)?)?;
